fix(markdown): update imports and remove unused code in Markdown components
