+++ src/components/ui/form/index.ts
@@ -4,8 +4,6 @@ export * from './Autocomplete'
 export * from './Input'
 export * from './InputPlaceholder'
 export * from './MarkdownEditor'
-export * from './MarkdownRenderer'
 export * from './MultiSelect'
 export * from './SelectInput'
 export * from './Toggle'
-export * from './TranslatableMarkdown'
