+++ src/components/ui/index.ts
@@ -52,5 +52,6 @@ export * from './VoteButtons'
 export * from './charts'
 export * from './form'
 export * from './image-selectors'
+export * from './markdown'
 export * from './modals'
 export * from './table-buttons'
